Simplify note_commit
